@@ -11,7 +11,7 @@
 		}
     ],
     "require": {
-		"php": ">=5.3.0"
+		"php": ">=5.4.0"
     },
  "autoload": {
         "psr-4": {
